+++ src/preload/index.ts
@@ -59,6 +59,7 @@ const validInvokeChannels = [
   // Custom Line Groups
   'getCustomLineGroupsConfig',
   'setCustomLineGroupsConfig',
+  'checkPortOccupied',
   // File
   'getFileStr',
   'setFileStr',